Undo last 'fix', since it was not the proper one.

+++ lib/hostip6.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: hostip6.c,v 1.23 2005-12-12 18:40:18 yangtse Exp $
+ * $Id: hostip6.c,v 1.24 2005-12-13 13:50:22 yangtse Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -143,9 +143,10 @@ int curl_dogetnameinfo(const struct sockaddr *sa, socklen_t salen,
                        char *serv, size_t servlen, int flags,
                        int line, const char *source)
 {
-  int res = (getnameinfo)(sa, (size_t)salen, 
-                          host, hostlen, serv, servlen, 
-                          (unsigned int)flags);
+  int res = (getnameinfo)(sa, salen, 
+                          host, hostlen, 
+                          serv, servlen, 
+                          flags); 
   if(0 == res) {
     /* success */
     if(logfile)
